LayoutTests/media/media-continues-playing-after-replace-source-expected.txt

Test that media keeps playing when the source element is replaced.

EVENT(canplaythrough)
EXPECTED (testElement.currentTime == '0') OK
EVENT(playing)
EXPECTED (testElement.currentTime > '0') OK
Replacing the media's source element:
EXPECTED (testElement.currentTime == '0') OK
EXPECTED (testElement.currentTime > '0') OK
END OF TEST
